+++ src/js/controllers/main/main.module.js
@@ -1,0 +1,5 @@
+import controller from "./main.controller";
+
+angular
+.module("uploader.main", [])
+.controller("MainController", controller);