+++ src/js/services/api/api.module.js
@@ -1,0 +1,5 @@
+import service from "./api.service";
+
+angular
+.module("uploader.api", [])
+.factory("API", service);