+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+### 0.2.2
+
+**Minor**
+
+- none
+
+**Patch**
+
+- split angular files a little more
+- renamed Ajax and MainCtrl to API and MainController
+- general John Papa styleguide stuff
+
 ### 0.2.1
 
 **Minor**